Updated the methods of the SphinxProperties object that retrieves property values of instances. The current policy is as follows. For example, when the:
getInt(String instanceName, String propertyName, int defaultValue);

is called, it first checks if there is a property called:

"<instanceName>;<propertyName>"

if yes, return the integer value. If no, it checks for the property:

"<propertyName>"

if it exists, return the integer. If it doesn't exist, return the default value.

This is the same for the getString, getFloat, getDouble, and getBoolean
methods that deal with properties of instances.