Previously, the WordPruningBreadthFirstSearchManager would get the an array of classes representing the type and order of states in the SearchGraph.  The WPBFSM used this class to construct an active list manager that maintains an active list for each class type. The ActiveListManager used a hash map to map a particular class to the appropriate active list. This mapping (according to the jfluid profile) accounted for around 5% of execution time.
This change:
a) makes the interface simpler, and
b) eliminates the need for the hash table lookup, replacing it with a simple
index.


The interface changes are as follows:

class SearchGraph:

eliminate method 'Class[] getSearchStateOrder()'

add method 'int getNumStateOrder()' returns the number of possible state
orders

To SearchState add:

   'int getOrder()' - returns the state order - 0 is furthest from emitting
state,


The ActiveListManager is able to use the index returned by
state.getOrder()  to index into an array of active lists as opposed to having
to use a hash map.

Also, this gives the linguist the ability to classify two different state
types as being at the same rank order, reducing the number of active lists
needed for the search.